test: 의존성 추가 및 마이그레이션에 따른 어노테이션 변경
